Add C++ solution for Number_Spiral

+++ Introductory_Problems/Number_Spiral/Number_Spiral.cpp
@@ -21,9 +21,9 @@ int main(){
             long long col = (long long)(x + rem - 2)*(x + rem - 1) + x; 
             
             if (x%2 == 0){
-                num = col + (y - 1); 
+                num = (long long)(col + (y - 1)); 
             }
-            else num = col - (y - 1); 
+            else num = (long long)(col - (y - 1)); 
             //cout<<col<<" "; 
         }
         else{
@@ -31,9 +31,9 @@ int main(){
             long long row = (long long)(y - rem - 1)*(y - rem) + y; 
             
             if (y%2 == 0){
-                num = row - (x - 1); 
+                num = (long long)(row - (x - 1)); 
             }
-            else num = row + (x - 1); 
+            else num = (long long)(row + (x - 1)); 
             //cout<<col<<" "; 
         }
         cout<<num<<endl;
